+++ .gitignore
@@ -4,3 +4,4 @@ punchout-ai/punchout_ai/envs/__pycache__/**/*
 .vscode/**/*
 BizHawk/**/*
 !BizHawk/BizHawk.Client.EmuHawk/tools/PunchoutBot/**/*
+classes/__pycache__/**/*
